refactored XNAT CS deploy build
